+++ src/serialization/json_object.h
@@ -225,6 +225,15 @@ void fromJsonValue(const rapidjson::Value& val, cryptonote::txin_to_key& txin);
 void toJsonValue(rapidjson::Writer<epee::byte_stream>& dest, const cryptonote::txin_stake_claim& txin);
 void fromJsonValue(const rapidjson::Value& val, cryptonote::txin_stake_claim& txin);
 
+void toJsonValue(rapidjson::Writer<epee::byte_stream>& dest, const cryptonote::archival_leaf_bytes& leaf);
+void fromJsonValue(const rapidjson::Value& val, cryptonote::archival_leaf_bytes& leaf);
+
+void toJsonValue(rapidjson::Writer<epee::byte_stream>& dest, const cryptonote::archival_segment_path_opening& path);
+void fromJsonValue(const rapidjson::Value& val, cryptonote::archival_segment_path_opening& path);
+
+void toJsonValue(rapidjson::Writer<epee::byte_stream>& dest, const cryptonote::txin_archival_serve_credit_response& txin);
+void fromJsonValue(const rapidjson::Value& val, cryptonote::txin_archival_serve_credit_response& txin);
+
 void toJsonValue(rapidjson::Writer<epee::byte_stream>& dest, const cryptonote::txout_target_v& txout);
 void fromJsonValue(const rapidjson::Value& val, cryptonote::txout_target_v& txout);